+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for defining functions.
 - Shorthand assignment operators.
+- Debug statement.
 
 ### Changed